# Faster bits and bernoulli