@thegodone 2024-10-28
Yes I see that bernoulli is bad top on torch but only on cpu, I made a issue last week.